feat: major pub fixes and state management for subcat_dropdown

+++ lib/widgets/dots_indicator_widget.dart
@@ -10,7 +10,8 @@ class DotsIndicatorWidget extends StatelessWidget {
     required this.dotscount,
   }) : super(key: key);
 
-  final double? position;
+  final int? position;
+  // final double? position;
   final int dotscount;
 
   @override
